RavenDB-9324 - PerformanceHints.TooLongRequestThresholdInSec should be configurable per database also, not only per server - performance hints should be persisted in database.NotificationCenter, not server.NotificationCenter - fixed polarization
